backend/requirements/base.txt
# Django and Extensions
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
django-extensions>=3.2,<3.3
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<2.10
pgvector>=0.2,<0.3

# Redis
redis>=5.0,<5.1

# AWS
boto3>=1.34,<1.35
django-storages>=1.14,<1.15

# Document Processing
PyPDF2>=3.0,<3.1
pdfplumber>=0.10,<0.11
python-magic>=0.4,<0.5

# Text Processing and Embeddings
langchain>=0.1,<0.2
langchain-openai>=0.0.5,<0.1
langchain-community>=0.0.10,<0.1
tiktoken>=0.5,<0.6
sentence-transformers>=2.3,<2.4

# LLM Providers
openai>=1.10,<1.11
anthropic>=0.8,<0.9

# Authentication and Security
pyjwt>=2.8,<2.9
cryptography>=42.0,<43.0

# API Documentation
drf-spectacular>=0.27,<0.28

# Utilities
requests>=2.31,<2.32
python-dateutil>=2.8,<2.9
pytz>=2024.1
whitenoise>=6.6,<6.7
